ability to set disk sizes in GCP
GCP by default sizes disks based on the OS image, and they are quite
tiny. This gives us good default breathing room, and the ability to
get a larger size if needed.

This just increases the size of the boot partition... if we want to do
anything fancy with NVME or whatever, that will have to be handled
elsewhere.

What I want to avoid is the situation where we stand up some default
infrastructure, start getting a demo or whatever going and then late
in the process find out the disk is full and we need to start over.